# Extend Approximation Framework to other numeric types