--- a/server/src/main/java/org/opensearch/index/mapper/NumberFieldMapper.java
+++ b/server/src/main/java/org/opensearch/index/mapper/NumberFieldMapper.java
@@ -356,23 +356,50 @@ public class NumberFieldMapper extends ParametrizedFieldMapper {
                     }
                     u = HalfFloatPoint.nextDown(u);
                 }
-                if (isSearchable && hasDocValues) {
-                    Query query = HalfFloatPoint.newRangeQuery(field, l, u);
-                    Query dvQuery = SortedNumericDocValuesField.newSlowRangeQuery(
+
+                Query dvQuery = hasDocValues
+                    ? SortedNumericDocValuesField.newSlowRangeQuery(
                         field,
                         HalfFloatPoint.halfFloatToSortableShort(l),
                         HalfFloatPoint.halfFloatToSortableShort(u)
+                    )
+                    : null;
+                if (isSearchable) {
+                    Query pointRangeQuery = HalfFloatPoint.newRangeQuery(field, l, u);
+                    Query query;
+                    if (dvQuery != null) {
+                        query = new IndexOrDocValuesQuery(pointRangeQuery, dvQuery);
+                        if (context.indexSortedOnField(field)) {
+                            query = new IndexSortSortedNumericDocValuesRangeQuery(
+                                field,
+                                HalfFloatPoint.halfFloatToSortableShort(l),
+                                HalfFloatPoint.halfFloatToSortableShort(u),
+                                query
+                            );
+                        }
+                    } else {
+                        query = pointRangeQuery;
+                    }
+                    return new ApproximateScoreQuery(
+                        query,
+                        new ApproximatePointRangeQuery(
+                            field,
+                            NumberType.HALF_FLOAT.encodePoint(l),
+                            NumberType.HALF_FLOAT.encodePoint(u),
+                            1,
+                            ApproximatePointRangeQuery.HALF_FLOAT_FORMAT
+                        )
                     );
-                    return new IndexOrDocValuesQuery(query, dvQuery);
                 }
-                if (hasDocValues) {
-                    return SortedNumericDocValuesField.newSlowRangeQuery(
+                if (context.indexSortedOnField(field)) {
+                    dvQuery = new IndexSortSortedNumericDocValuesRangeQuery(
                         field,
                         HalfFloatPoint.halfFloatToSortableShort(l),
-                        HalfFloatPoint.halfFloatToSortableShort(u)
+                        HalfFloatPoint.halfFloatToSortableShort(u),
+                        dvQuery
                     );
                 }
-                return HalfFloatPoint.newRangeQuery(field, l, u);
+                return dvQuery;
             }
 
             @Override
@@ -503,14 +530,28 @@ public class NumberFieldMapper extends ParametrizedFieldMapper {
                         u = FloatPoint.nextDown(u);
                     }
                 }
-                if (isSearchable && hasDocValues) {
+                if (isSearchable) {
                     Query query = FloatPoint.newRangeQuery(field, l, u);
-                    Query dvQuery = SortedNumericDocValuesField.newSlowRangeQuery(
-                        field,
-                        NumericUtils.floatToSortableInt(l),
-                        NumericUtils.floatToSortableInt(u)
+
+                    if (hasDocValues) {
+                        Query dvQuery = SortedNumericDocValuesField.newSlowRangeQuery(
+                            field,
+                            NumericUtils.floatToSortableInt(l),
+                            NumericUtils.floatToSortableInt(u)
+                        );
+                        query = new IndexOrDocValuesQuery(query, dvQuery);
+                    }
+
+                    return new ApproximateScoreQuery(
+                        query,
+                        new ApproximatePointRangeQuery(
+                            field,
+                            FloatPoint.pack(new float[] { l }).bytes,
+                            FloatPoint.pack(new float[] { u }).bytes,
+                            1,
+                            ApproximatePointRangeQuery.FLOAT_FORMAT
+                        )
                     );
-                    return new IndexOrDocValuesQuery(query, dvQuery);
                 }
                 if (hasDocValues) {
                     return SortedNumericDocValuesField.newSlowRangeQuery(
@@ -628,23 +669,49 @@ public class NumberFieldMapper extends ParametrizedFieldMapper {
                 QueryShardContext context
             ) {
                 return doubleRangeQuery(lowerTerm, upperTerm, includeLower, includeUpper, (l, u) -> {
-                    if (isSearchable && hasDocValues) {
-                        Query query = DoublePoint.newRangeQuery(field, l, u);
-                        Query dvQuery = SortedNumericDocValuesField.newSlowRangeQuery(
+                    Query dvQuery = hasDocValues
+                        ? SortedNumericDocValuesField.newSlowRangeQuery(
                             field,
                             NumericUtils.doubleToSortableLong(l),
                             NumericUtils.doubleToSortableLong(u)
+                        )
+                        : null;
+                    if (isSearchable) {
+                        Query pointRangeQuery = DoublePoint.newRangeQuery(field, l, u);
+                        Query query;
+                        if (dvQuery != null) {
+                            query = new IndexOrDocValuesQuery(pointRangeQuery, dvQuery);
+                            if (context.indexSortedOnField(field)) {
+                                query = new IndexSortSortedNumericDocValuesRangeQuery(
+                                    field,
+                                    NumericUtils.doubleToSortableLong(l),
+                                    NumericUtils.doubleToSortableLong(u),
+                                    query
+                                );
+                            }
+                        } else {
+                            query = pointRangeQuery;
+                        }
+                        return new ApproximateScoreQuery(
+                            query,
+                            new ApproximatePointRangeQuery(
+                                field,
+                                DoublePoint.pack(new double[] { l }).bytes,
+                                DoublePoint.pack(new double[] { u }).bytes,
+                                1,
+                                ApproximatePointRangeQuery.DOUBLE_FORMAT
+                            )
                         );
-                        return new IndexOrDocValuesQuery(query, dvQuery);
                     }
-                    if (hasDocValues) {
-                        return SortedNumericDocValuesField.newSlowRangeQuery(
+                    if (context.indexSortedOnField(field)) {
+                        dvQuery = new IndexSortSortedNumericDocValuesRangeQuery(
                             field,
                             NumericUtils.doubleToSortableLong(l),
-                            NumericUtils.doubleToSortableLong(u)
+                            NumericUtils.doubleToSortableLong(u),
+                            dvQuery
                         );
                     }
-                    return DoublePoint.newRangeQuery(field, l, u);
+                    return dvQuery;
                 });
             }
 
@@ -988,23 +1055,33 @@ public class NumberFieldMapper extends ParametrizedFieldMapper {
                         --u;
                     }
                 }
-                if (isSearchable && hasDocValues) {
-                    Query query = IntPoint.newRangeQuery(field, l, u);
-                    Query dvQuery = SortedNumericDocValuesField.newSlowRangeQuery(field, l, u);
-                    query = new IndexOrDocValuesQuery(query, dvQuery);
-                    if (context.indexSortedOnField(field)) {
-                        query = new IndexSortSortedNumericDocValuesRangeQuery(field, l, u, query);
+                Query dvQuery = hasDocValues ? SortedNumericDocValuesField.newSlowRangeQuery(field, l, u) : null;
+                if (isSearchable) {
+                    Query pointRangeQuery = IntPoint.newRangeQuery(field, l, u);
+                    Query query;
+                    if (dvQuery != null) {
+                        query = new IndexOrDocValuesQuery(pointRangeQuery, dvQuery);
+                        if (context.indexSortedOnField(field)) {
+                            query = new IndexSortSortedNumericDocValuesRangeQuery(field, l, u, query);
+                        }
+                    } else {
+                        query = pointRangeQuery;
                     }
-                    return query;
+                    return new ApproximateScoreQuery(
+                        query,
+                        new ApproximatePointRangeQuery(
+                            field,
+                            IntPoint.pack(new int[] { l }).bytes,
+                            IntPoint.pack(new int[] { u }).bytes,
+                            1,
+                            ApproximatePointRangeQuery.INT_FORMAT
+                        )
+                    );
                 }
-                if (hasDocValues) {
-                    Query query = SortedNumericDocValuesField.newSlowRangeQuery(field, l, u);
-                    if (context.indexSortedOnField(field)) {
-                        query = new IndexSortSortedNumericDocValuesRangeQuery(field, l, u, query);
-                    }
-                    return query;
+                if (context.indexSortedOnField(field)) {
+                    dvQuery = new IndexSortSortedNumericDocValuesRangeQuery(field, l, u, dvQuery);
                 }
-                return IntPoint.newRangeQuery(field, l, u);
+                return dvQuery;
             }
 
             @Override
@@ -1136,11 +1213,10 @@ public class NumberFieldMapper extends ParametrizedFieldMapper {
                                 field,
                                 LongPoint.pack(new long[] { l }).bytes,
                                 LongPoint.pack(new long[] { u }).bytes,
-                                new long[] { l }.length,
+                                1,
                                 ApproximatePointRangeQuery.LONG_FORMAT
                             )
                         );
-
                     }
                     if (context.indexSortedOnField(field)) {
                         dvQuery = new IndexSortSortedNumericDocValuesRangeQuery(field, l, u, dvQuery);
@@ -1257,10 +1333,22 @@ public class NumberFieldMapper extends ParametrizedFieldMapper {
                 QueryShardContext context
             ) {
                 return unsignedLongRangeQuery(lowerTerm, upperTerm, includeLower, includeUpper, (l, u) -> {
-                    if (isSearchable && hasDocValues) {
+                    if (isSearchable) {
                         Query query = BigIntegerPoint.newRangeQuery(field, l, u);
-                        Query dvQuery = SortedUnsignedLongDocValuesRangeQuery.newSlowRangeQuery(field, l, u);
-                        return new IndexOrDocValuesQuery(query, dvQuery);
+                        if (hasDocValues) {
+                            Query dvQuery = SortedUnsignedLongDocValuesRangeQuery.newSlowRangeQuery(field, l, u);
+                            query = new IndexOrDocValuesQuery(query, dvQuery);
+                        }
+                        return new ApproximateScoreQuery(
+                            query,
+                            new ApproximatePointRangeQuery(
+                                field,
+                                NumberType.UNSIGNED_LONG.encodePoint(l),
+                                NumberType.UNSIGNED_LONG.encodePoint(u),
+                                1,
+                                ApproximatePointRangeQuery.UNSIGNED_LONG_FORMAT
+                            )
+                        );
                     }
                     if (hasDocValues) {
                         return SortedUnsignedLongDocValuesRangeQuery.newSlowRangeQuery(field, l, u);