… CHANGELOG.md
… server/src/test/java/org/opensearch/index/query/RangeQueryBuilderTests.java
… server/src/test/java/org/opensearch/search/approximate/ApproximatePointRangeQueryTests.java