<!--  Thanks for sending a pull request, here are some tips:

1. If this is a fix for an undisclosed security vulnerability, please STOP. All security vulnerability reporting and fixes should be done as per our security policy https://github.com/opensearch-project/OpenSearch/security/policy
2. If this is your first time, please read our contributor guidelines: https://github.com/opensearch-project/OpenSearch/blob/main/CONTRIBUTING.md and developer guide https://github.com/opensearch-project/OpenSearch/blob/main/DEVELOPER_GUIDE.md
3. Ensure you have added or ran the appropriate tests for your PR: https://github.com/opensearch-project/OpenSearch/blob/main/TESTING.md
-->

### Description
- Today `LONG` already goes through the `ApproximateScoreQuery`.

- Extend `HALF_FLOAT`, `FLOAT`, `DOUBLE`, `INTEGER`, `UNSIGNED_LONG` to use `ApproximateScoreQuery`. `BYTE` and `SHORT` uses `INTEGER`.

- Updated the `rangeQuery` methods to check for `indexSortedOnField`. This is exception for `UNSIGNED_LONG` as coming from https://docs.opensearch.org/docs/latest/field-types/supported-field-types/unsigned-long/ index sort is not supported for `UNSIGNED_LONG`. 
    - Seeing this https://github.com/apache/lucene/blob/main/lucene/core/src/java/org/apache/lucene/search/SortField.java I can see not just `INTEGER` and `LONG` there are other types as well (**still confirming and investigating more on this**).

- Since `HALF_FLOAT` and `UNSIGNED_LONG` are still in sandbox usage of `.pack` is not supported. Here is the PR i have created in Lucene https://github.com/apache/lucene/pull/14784. Used existing `encode` methods.


### Related Issues
https://github.com/opensearch-project/OpenSearch/issues/14406 and part of https://github.com/opensearch-project/OpenSearch/issues/18334


### Check List
- [ ] Functionality includes testing.
- [ ] API changes companion pull request [created](https://github.com/opensearch-project/opensearch-api-specification/blob/main/DEVELOPER_GUIDE.md), if applicable.
- [ ] Public documentation issue/PR [created](https://github.com/opensearch-project/documentation-website/issues/new/choose), if applicable.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.
For more information on following Developer Certificate of Origin and signing off your commits, please check [here](https://github.com/opensearch-project/OpenSearch/blob/main/CONTRIBUTING.md#developer-certificate-of-origin).