@prudhvigodithi 2025-06-24
@kkewwei can you please take a look at this PR? I have added some parameterized tests updating `ApproximatePointRangeQueryTests` which would run for all numeric types, this way we dont have to re-write all tests again for each numeric type.