@kkewwei 2025-07-01
> @kkewwei for [#18530 (comment)](https://github.com/opensearch-project/OpenSearch/pull/18530#discussion_r2158442607) Initially that was my plan as well, but looking deep each NumericType rangeQuery method has its own clauses and some logic involved. Example today the `indexSortedOnField` to use `IndexSortSortedNumericDocValuesRangeQuery` is not part of all range methods however it should be true only for `unsigned_long` as per https://docs.opensearch.org/docs/latest/field-types/supported-field-types/unsigned-long/ (`An unsigned_long field cannot be used as an index sort field (in the sort.field index setting).`) Diving little deep here.
> 
> I'm focusing on adding tests to `ApproximatePointRangeQueryTests` for all numeric types, exploring OpenSearch testing `ParametersFactory` to that we can re-use the existing tests for all numeric types. Once we are good and ensure performance improved and results are same as expected we can update the `NumberFieldMapper` with more optimized range query.

I concur with your revision. Notably, most rangeQuery operations currently exhibit uniform schema structures,  looking forward to further simplification of the codebase in the upcoming iterations.