@prudhvigodithi 2025-07-03
From above benchmark comparison results https://github.com/opensearch-project/OpenSearch/pull/18530#issuecomment-3029751497 (`nyc_taxis`) https://github.com/opensearch-project/OpenSearch/pull/18530#issuecomment-3029809877 (`http_logs`) seen good improvement with sorts, range and match_all query shapes, for some of them like `desc_sort_tip_amount` seen 60% improvement of p90 from 26.3ms to 10.5ms, for `range_size` seen 80% improvement of p90 from 50.3ms to 8.8ms, for `range_with_asc_sort` seen more than 80% improvement of p90 when from 290ms to 22ms and seen same with `range_with_desc_sort` seen more than 80% improvement of p90 went down from 383.4ms 27.2ms.